Update faq.md
Scavenge levelling add: everyone asks the "how to start scavenge????" question so much I have to add this. I already did in tips but double is better.
Ammo crafting and mining skill benefits add: ratio isn't exactly the correct word to use here because it makes it look like the ratio is the same, i.e. you have to spend 10 military ammo to craft 1 AMC, so by level 100 you need 100 to craft 10. Changed wording so that it uses *same amount* rather than *same ratio* (it fixes the number to be 10 for example but increases the output crafted)

also some more changes on grammar and a few hanging threads that are just there. haha some changes xd