Update for the german lang for reading quality
Hello, Here is a update for the German translation with some fitting words and grammatical error removal.
Quest - Suche are non-related. "Suche" means search, a better fitting word would be "Aufträge" because the quest you see and complete would be called Aufträge in Germany.
Chat - Plaudern is just out of time and isnt even fitting with the theme of the Game. "Kommunikation" would fit better with the meaning of "chat" and the theme of the game. Plaudern is just a really specific word for talking that we Germans don't use often.
Kosmetischer Cyberwear - grammatical error, remove the r = Kosmetische Cyberwear sounds so much better now

Upper - lower - boots.
So hear me out first. "Rüstung" "Stiefel" "Helmet" are just 1on1 translation. Those aren't really fitting for the different kind of equipments. Stiefel are special kind of boots, if you get flipflop boots it contradicts itself. Use "Schuhe" instead. to better fit the combination of Helm and Schuhe, Jacke and Hose would fit right in and its still clear what kind of armor parts those are. It would also be more comfortable for reading.

Spiele MEGA Pachinko would be a better 1 to 1 translation.
Street Cred - Straßen Glaube, grammatical error.
"Give Street Cred" - so its german and fitting 
"Spielanleitung" would fit better but you could always just use "tutorial". Some lines above there is the same translation of it with tutorial (profile tab).

I discussed a lot of those translation errors with other Germans in game and we all agree that some of the old worlds could not be read right bc there just not fitting for the game and there purpose.