Update tips.txt
That tip is just peak confusion and caused confusion in global. If it was supposed to mean modules it can't, as base stats of a piece of gear it can't, and as cali'd gear it makes no sense to post a tip specifically about cd only instead of cc, eva, or stun.